- Missing wads message box will now display more accurate error/warning messages. - Introduced Message system that may be used to normalize the way messages are passed and interpreted throughout the application. - Removed now obsolete MessageResult class. - Messages for missing binaries should now be more accurate. - Decreased the default number of server query attempts to 3 (if it's not gonna respond then it's not gonna respond).

+++ src/odamex/odamexserver.cpp
@@ -107,7 +107,7 @@ Server::Response OdamexServer::readRequest(QByteArray &data)
 		else if(cvarName == "sv_gametype")
 		{
 			unsigned int mode = cvarValue.toUInt();
-			if(mode < plugin()->pInterface->gameModes()->size())
+			if(mode < (unsigned)plugin()->pInterface->gameModes()->size())
 				currentGameMode = (*plugin()->pInterface->gameModes())[cvarValue.toUInt()];
 		}
 	}
